+++ crates/alkahest-renderer/src/renderer/shadows.rs
@@ -4,12 +4,13 @@ use alkahest_data::{
     technique::StateSelection,
     tfx::{TfxRenderStage, TfxShaderStage},
 };
-use bevy_ecs::entity::Entity;
+use bevy_ecs::{entity::Entity, system::RunSystemOnce, world::Ref};
 
 use crate::{
     ecs::{
         render::light::{ShadowGenerationMode, ShadowMapRenderer},
         transform::Transform,
+        visibility::{reset_view_visibility_system, ViewVisibility, VisibilityHelper},
         Scene,
     },
     gpu_event,
@@ -34,11 +35,15 @@ impl Renderer {
         self.gpu.flush_states();
 
         let mut shadow_renderers = vec![];
-        for (e, shadow) in scene
-            .query::<(Entity, &mut ShadowMapRenderer)>()
+        for (e, shadow, view_vis) in scene
+            .query::<(Entity, &mut ShadowMapRenderer, Option<&ViewVisibility>)>()
             .iter(scene)
         {
-            shadow_renderers.push((e, shadow.last_update));
+            // TODO(cohae): view visibility might change a bit, since shadow maps are technically views as well
+            // Only update shadow maps for visible lights
+            if view_vis.is_visible(0) || !self.data.lock().asset_manager.is_idle() {
+                shadow_renderers.push((e, shadow.last_update));
+            }
         }
 
         shadow_renderers.sort_by_key(|(_, last_update)| *last_update);
@@ -56,7 +61,7 @@ impl Renderer {
                 .shadowmap_vs_t2
                 .bind(&self.gpu, 2, TfxShaderStage::Vertex);
 
-            self.bind_view(&*shadow);
+            self.bind_view(&*shadow, e.index() as usize);
 
             if shadow.stationary_needs_update {
                 self.pocus().active_shadow_generation_mode = ShadowGenerationMode::StationaryOnly;