+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ### Added
 
+- Cull static/dynamic geometry outside of the view frustum by @cohaereo
 - Write tracing events to alkahest.log by @cohaereo
 
 ### Changed